More work on profile page

+++ frontend/src/stores/authStore.ts
@@ -23,21 +23,42 @@ export const useAuthStore = defineStore("auth", () => {
         credentials: "include",
       });
 
+      if (loginResponse.status === 401) {
+        localStorage.clear();
+
+        isLoggedIn.value = false;
+        profileStore.profile = null;
+
+        router.push("/");
+        return;
+      }
+
       isLoggedIn.value = loginResponse.status === 200;
     } catch (err) {
-      // TODO: handle error
+      // TODO: handle error (use toast system)
       console.error(err);
       isLoggedIn.value = false;
     }
   }
 
-  function logout() {
-    localStorage.clear();
+  async function logout() {
+    try {
+      const accessToken = localStorage.getItem("access_token");
+      await fetch(`${authApiBaseUrl}/logout`, {
+        method: "POST",
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+        credentials: "include",
+      });
+    } finally {
+      localStorage.clear();
 
-    isLoggedIn.value = false;
-    profileStore.profile = null;
+      isLoggedIn.value = false;
+      profileStore.profile = null;
 
-    router.push("/");
+      router.push("/");
+    }
   }
 
   function validateToken() {
